DChargedTrack::Get_Charge now returns an int, DNeutralParticle::Get_BestPhoton and ::Get_BestNeutron are replaced with ::Get_Hypothesis, Get_BestFOM in both classes no longer assumes the objects are stored in order of FOM.

+++ src/libraries/PID/DChargedTrack.h
@@ -21,7 +21,7 @@ class DChargedTrack:public jana::JObject{
 
 		vector<const DChargedTrackHypothesis*> dChargedTrackHypotheses;
 
-		double Get_Charge(void) const;
+		int Get_Charge(void) const;
 		const DChargedTrackHypothesis* Get_Hypothesis(Particle_t locPID) const;
 		const DChargedTrackHypothesis* Get_BestFOM(void) const;
